backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
opencv-python==4.10.0.84
reportlab==4.2.2
websockets==12.0
python-dotenv==1.0.0
openai
google-generativeai
google-cloud-texttospeech
gTTS
